+++ src/helpers.php
@@ -1,0 +1,18 @@
+<?php
+
+if ( ! function_exists('backpack_url') ) {
+    /**
+     * Appends the configured backpack prefix and returns
+     * the URL using the standard Laravel helpers.
+     *
+     * @param $path
+     *
+     * @return string
+     */
+    function backpack_url( $path = null )
+    {
+        $path = ! $path || (substr($path, 1, 1) == '/') ? $path : '/' . $path;
+        return url(config('backpack.base.route_prefix', 'admin') . $path);
+    }
+
+}