+++ src/resources/views/auth/register.blade.php
@@ -8,7 +8,7 @@
                     <div class="box-title">{{ trans('backpack::base.register') }}</div>
                 </div>
                 <div class="box-body">
-                    <form class="form-horizontal" role="form" method="POST" action="{{ url(config('backpack.base.route_prefix', 'admin').'/register') }}">
+                    <form class="form-horizontal" role="form" method="POST" action="{{ backpack_url('register') }}">
                         {!! csrf_field() !!}
 
                         <div class="form-group{{ $errors->has('name') ? ' has-error' : '' }}">